refactor: change package structure from Nuxt Basic to Dix Convention
