+++ component/Icon/Component.vue
