+++ component/SearchBar.vue
