+++ util/createRoom.ts
@@ -1,5 +1,6 @@
 import { addDoc, collection, serverTimestamp } from "firebase/firestore";
-import { db } from "~/utils/firebase";
+
+import { db } from "~/util/firebase";
 
 export async function createChatRoom(name: string, participants: string[]) {
   const roomRef = await addDoc(collection(db, "chatRooms"), {
